requirements.txt
streamlit
fastapi
uvicorn
python-dotenv  
rank_bm25
torch
sentence-transformers
pyvi